remove grid
